Flight ready no callsign
GSV, GSA and GLL sentences removed

+++ aprs.cpp
@@ -34,7 +34,7 @@ extern SoftwareSerial softdebug;
 #endif
 
 // Module functions
-long meters_to_feet(long m)
+float meters_to_feet(float m)
 {
   // 10000 ft = 3048 m
   return m / 0.3048;
@@ -75,7 +75,7 @@ void aprs_send()
   ax25_send_byte('/');                // and
   snprintf(temp, 4, "%03d", (int)(kmh_to_knots(gps_speed) + 0.5));
   ax25_send_string(temp);             // speed (knots)
-  ax25_send_string("/A=");            // Altitude (meters). Goes anywhere in the comment area
+  ax25_send_string("/A=");            // Altitude (feet). Goes anywhere in the comment area
   snprintf(temp, 7, "%06ld", (long)(meters_to_feet(gps_altitude) + 0.5));
   ax25_send_string(temp);
   ax25_send_string("/G");            // the number of milliseconds since the last valid gps position
